refactor: separate statistics data class from Inner Class into separate package
