+++ src/main/java/com/example/spinlog/statistics/dto/cache/CountsAndSums.java
@@ -1,0 +1,15 @@
+package com.example.spinlog.statistics.dto.cache;
+
+import java.util.Map;
+
+public record CountsAndSums(
+        Map<String, Object> sumsMap,
+        Map<String, Object> countsMap) {
+    @Override
+    public String toString() {
+        return "CountsAndSums{" +
+                "\n\tsumsMap=" + sumsMap +
+                "\n\tcountsMap=" + countsMap +
+                '}';
+    }
+}
