+++ src/main/java/com/example/spinlog/statistics/service/cache/scheduled/startup/GenderStatisticsCacheStartupService.java
@@ -4,6 +4,7 @@ import com.example.spinlog.global.cache.CacheHashRepository;
 import com.example.spinlog.statistics.service.StatisticsPeriodManager;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
 import com.example.spinlog.statistics.utils.StatisticsZeroPaddingUtils;
+import com.example.spinlog.statistics.dto.cache.AllStatisticsMap;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
@@ -15,7 +16,6 @@ import java.time.LocalDate;
 import static com.example.spinlog.article.entity.RegisterType.SAVE;
 import static com.example.spinlog.article.entity.RegisterType.SPEND;
 import static com.example.spinlog.statistics.service.StatisticsPeriodManager.*;
-import static com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService.AllStatisticsMap;
 import static com.example.spinlog.statistics.utils.CacheKeyNameUtils.*;
 
 @Component
