+++ src/test/java/com/example/spinlog/statistics/service/cache/GenderStatisticsCacheWriteServiceTest.java
@@ -2,6 +2,8 @@ package com.example.spinlog.statistics.service.cache;
 
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.global.cache.CacheHashRepository;
+import com.example.spinlog.statistics.dto.cache.AllStatisticsMap;
+import com.example.spinlog.statistics.dto.cache.CountsAndSums;
 import com.example.spinlog.util.MockCacheHashRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.ActiveProfiles;
